[WebXR] Add OffscreenCanvas support for makeXrCompatible
When an offscreen canvas doesn't have a placeholder canvas (such as
when new OffscreenCanvas(width, height) is called to create an offscreen
canvas in javascript), the corresponding C++ object is OffscreenCanvas,
not HTMLCanvasElement. This change obtains the XRSystem object from
the OffscreenCanvas so that MakeXrCompatible can be called on it.

Bug: 1154424
Change-Id: I4fb1c7005a8c3a41cc4ddcd457c81c4bec4464ab