Change getImageData and createImageData to default to 8-bit sRGB
In the initial implementation, if color setting information is missing
in getImageData/createImageData, it uses the information of canvas to
initialize the image data. In the spec, it says that it should use the
default setting of {colorSpace: srgb, storageFormat: uint8} instead. In
this cl, I updated the default behavior for the function getImageData
and createImageData.

Bug:1118182

Change-Id: I4fba4284028c060d5ac7ff80a294d95f8436ef62