[overflow] Fix negative margin compat issues.
Many sites use negative block-end margins to "pull" in the end padding.
(This is likely because FF doesn't support this end padding, and
"working" around this fact").

For webcompat this change allows negative margins for determining the
inflow bounds, but only in the scrollable direction.

While manual testing I found also that the simplified layout pass
had a issue of not calling AddChild correctly with all the necessary
information.

Bug: 1157299
Change-Id: I90f24301a7015915b0cd0a4446940ee562876dc0