[MSE][WebCodecs] Stubs new WebCodecEncodedChunkStreamParser
Adds a new "StreamParser" subtype for buffering WebCodecs encoded audio
or video chunks. Plumbs the initial decoder configuration (to later be
emitted during eventual first ProcessChunks() call on the new parser)
through ChunkDemuxer's AddId to a new instance of one of these parsers.

Includes a StreamParserFactory method to create the new kind of parser.

Updates the web_test to expect a SourceBuffer to be created for valid
WebCodecs config cases (instead of a QuotaExceededError exception as was
stubbed previously.)

Later changes will complete the implementation of
WebCodecsEncodedChunkStreamParser::ProcessChunks() and include more
testing of end-to-end buffering and playback.

BUG=1144908

Change-Id: Ifb5e88457cdbc8ccfe37d5ce2e6bee07fec13726