SXG + reporting: Make WPT tests work with NetworkIsolationKeys.
NetworkIsolationKeys include top frame origin and frame origin.  The SXG
The WPT SXG reporting tests load an SXG in a cross-site iframe, but
learn reporting information in the top-level browsing context, so it
had a different NIK.

This CL fixes those tests to learn reporting information in an iframe,
so they have the same NetworkIsolationKey as the signed exchange. With
this CL, all SXG WPT tests should now pass with reporting data
partitioned by NetworkIsolationKey.

Bug: 1156951
Change-Id: I0f8aa3bc51a276a71c28e92a056e709721bffb04