Correct painting order of SVG mask and mask-based clip-path
In paint properties, if both mask and mask-based clip-path exist, we
make the ClipPathMask effect node as a child of Mask node, which
requires us to paint mask before clip-path mask. The latter applies on
the former to create an intersection of the masks, then the intersection
will be applied to the masked content.

Previously the clip-path mask was painted before the mask, and the
former applies to a transparent backdrop which had no effect.

Bug: 1155384
Change-Id: Ic40c8254193c141a286c4b323515fabf8a76e11a